+++ src/test/java/com/expedia/www/haystack/metrics/ServoToInfluxDbViaGraphiteNamingConventionTest.java
@@ -33,6 +33,7 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.Random;
 
+import static com.expedia.www.haystack.metrics.MetricObjects.METRIC_GROUP_BUCKETS;
 import static com.expedia.www.haystack.metrics.MetricObjects.TAG_KEY_APPLICATION;
 import static com.expedia.www.haystack.metrics.MetricObjects.TAG_KEY_FULLY_QUALIFIED_CLASS_NAME;
 import static com.expedia.www.haystack.metrics.MetricObjects.TAG_KEY_LINE_NUMBER;
@@ -40,7 +41,8 @@ import static com.expedia.www.haystack.metrics.MetricObjects.TAG_KEY_METRIC_GROU
 import static com.expedia.www.haystack.metrics.ServoToInfluxDbViaGraphiteNamingConvention.METRIC_FORMAT_6_ARGS;
 import static com.expedia.www.haystack.metrics.ServoToInfluxDbViaGraphiteNamingConvention.METRIC_FORMAT_7_ARGS;
 import static com.expedia.www.haystack.metrics.ServoToInfluxDbViaGraphiteNamingConvention.MISSING_TAG;
-import static com.expedia.www.haystack.metrics.ServoToInfluxDbViaGraphiteNamingConvention.STATISTIC_TAG_NAME;
+import static com.expedia.www.haystack.metrics.ServoToInfluxDbViaGraphiteNamingConvention.TAG_KEY_SERVO_BUCKET;
+import static com.expedia.www.haystack.metrics.ServoToInfluxDbViaGraphiteNamingConvention.TAG_KEY_STATISTIC;
 import static com.expedia.www.haystack.metrics.MetricObjects.TAG_KEY_CLASS;
 import static com.expedia.www.haystack.metrics.MetricObjects.TAG_KEY_SUBSYSTEM;
 import static org.junit.Assert.assertEquals;
@@ -48,19 +50,20 @@ import static org.mockito.Mockito.verifyNoMoreInteractions;
 
 @RunWith(MockitoJUnitRunner.class)
 public class ServoToInfluxDbViaGraphiteNamingConventionTest {
-    private final static Random RANDOM = new Random();
-    private final static String METRIC_NAME = RANDOM.nextLong() + "METRIC_NAME";
-    private final static String METRIC_GROUP = RANDOM.nextLong() + "METRIC_GROUP";
-    private final static String SUBSYSTEM = RANDOM.nextLong() + "SUBSYSTEM";
-    private final static String APPLICATION = RANDOM.nextLong() + "APPLICATION";
-    private final static String FULLY_QUALIFIED_CLASS_NAME = RANDOM.nextLong() + "FULLY_QUALIFIED_CLASS_NAME";
-    private final static String CLASS = RANDOM.nextLong() + "CLASS";
-    private final static String LINE_NUMBER = Integer.toString(RANDOM.nextInt(Integer.MAX_VALUE));
-    private final static String TYPE = RANDOM.nextLong() + "TYPE";
-    private final static String STATISTIC = RANDOM.nextLong() + "STATISTIC";
+    private static final Random RANDOM = new Random();
+    private static final String METRIC_NAME = RANDOM.nextLong() + "METRIC_NAME";
+    private static final String METRIC_GROUP = RANDOM.nextLong() + "METRIC_GROUP";
+    private static final String SUBSYSTEM = RANDOM.nextLong() + "SUBSYSTEM";
+    private static final String APPLICATION = RANDOM.nextLong() + "APPLICATION";
+    private static final String BUCKET = RANDOM.nextLong() + "BUCKET";
+    private static final String FULLY_QUALIFIED_CLASS_NAME = RANDOM.nextLong() + "FULLY_QUALIFIED_CLASS_NAME";
+    private static final String CLASS = RANDOM.nextLong() + "CLASS";
+    private static final String LINE_NUMBER = Integer.toString(RANDOM.nextInt(Integer.MAX_VALUE));
+    private static final String TYPE = RANDOM.nextLong() + "TYPE";
+    private static final String STATISTIC = RANDOM.nextLong() + "STATISTIC";
     private static final String TYPE_STATISTIC = TYPE + '_' + STATISTIC;
-    private final static String LOCAL_HOST_NAME = "127.0.0.1";
-    private final static String LOCAL_HOST_NAME_CLEANED = LOCAL_HOST_NAME.replace(".", "_");
+    private static final String LOCAL_HOST_NAME = "127.0.0.1";
+    private static final String LOCAL_HOST_NAME_CLEANED = LOCAL_HOST_NAME.replace(".", "_");
 
     @Mock
     private InetAddress mockLocalHost;
@@ -113,7 +116,7 @@ public class ServoToInfluxDbViaGraphiteNamingConventionTest {
         tagList.add(Tags.newTag(TAG_KEY_APPLICATION, APPLICATION));
         tagList.add(Tags.newTag(TAG_KEY_CLASS, CLASS));
         tagList.add(Tags.newTag(DataSourceType.KEY, TYPE));
-        tagList.add(Tags.newTag(STATISTIC_TAG_NAME, STATISTIC));
+        tagList.add(Tags.newTag(TAG_KEY_STATISTIC, STATISTIC));
         final Metric metric = new Metric(METRIC_NAME, new BasicTagList(tagList), 0, 0);
 
         final String name = servoToInfluxDbViaGraphiteNamingConvention.getName(metric);
@@ -125,19 +128,33 @@ public class ServoToInfluxDbViaGraphiteNamingConventionTest {
 
     @Test
     public void testGetNameErrorCase() {
+        testGetNameCommonCode(METRIC_GROUP,
+                TAG_KEY_FULLY_QUALIFIED_CLASS_NAME, FULLY_QUALIFIED_CLASS_NAME,
+                TAG_KEY_LINE_NUMBER, LINE_NUMBER);
+    }
+
+    @Test
+    public void testGetNameBucketCase() {
+        testGetNameCommonCode(METRIC_GROUP_BUCKETS,
+                TAG_KEY_APPLICATION, APPLICATION,
+                TAG_KEY_SERVO_BUCKET, BUCKET);
+    }
+
+    private void testGetNameCommonCode(String metricGroup, String key1, String value1, String key2, String value2) {
         final List<Tag> tagList = new ArrayList<>(6);
-        tagList.add(Tags.newTag(TAG_KEY_METRIC_GROUP, METRIC_GROUP));
+        tagList.add(Tags.newTag(TAG_KEY_METRIC_GROUP, metricGroup));
         tagList.add(Tags.newTag(TAG_KEY_SUBSYSTEM, SUBSYSTEM));
-        tagList.add(Tags.newTag(TAG_KEY_FULLY_QUALIFIED_CLASS_NAME, FULLY_QUALIFIED_CLASS_NAME));
-        tagList.add(Tags.newTag(TAG_KEY_LINE_NUMBER, LINE_NUMBER));
+        tagList.add(Tags.newTag(key1, value1));
+        tagList.add(Tags.newTag(key2, value2));
         tagList.add(Tags.newTag(DataSourceType.KEY, TYPE));
-        tagList.add(Tags.newTag(STATISTIC_TAG_NAME, STATISTIC));
+        tagList.add(Tags.newTag(TAG_KEY_STATISTIC, STATISTIC));
         final Metric metric = new Metric(METRIC_NAME, new BasicTagList(tagList), 0, 0);
 
         final String name = servoToInfluxDbViaGraphiteNamingConvention.getName(metric);
 
-        final String expected = String.format(METRIC_FORMAT_7_ARGS, METRIC_GROUP, SUBSYSTEM, FULLY_QUALIFIED_CLASS_NAME,
-                LOCAL_HOST_NAME_CLEANED, LINE_NUMBER, METRIC_NAME, TYPE_STATISTIC);
+        final String expected = String.format(METRIC_FORMAT_7_ARGS, metricGroup, SUBSYSTEM,
+                value1, LOCAL_HOST_NAME_CLEANED, value2, METRIC_NAME, TYPE_STATISTIC);
         assertEquals(expected, name);
     }
+
 }